Allow typeName to be specified for table options
Adding the typeName avoids a lot of problems where any two tables with the same field names but different data types will conflict with each other, and cause all sorts of unclear warnings when generating Gatsby nodes.

`typeName` is specified on a table-by-table basis, and if not present the standard ‘Airtable’ is used as before.